Add about content, fix header visual bug, tighten up v-rhythm
There was a bug with the sticky headers, scroll context doesn't
appear to include borders. When passing off from one header to the next
in the posts list there was a little gap where the main header would
show through, yuk. So I'm using `<hr />`s in place of borders.

The placement of the titles on the tag pages and the titles of posts are
now the same. It has made the post titles maybe a little tighter than
I'd like.