Admit defeat for now and restart every hour
Some part of the static file server is leaking open files. Eventually
the ulimit is exceeded and the server stops being able to serve traffic.
It doesn't _exit_ however so I need to manually restart it whenever this
happens. I think I'm going to first try moving from
wai-middleware-static to wai-app-static (it was this initially) to see
if that alleviates the issue. I'm really not sure where the leak is
coming from though so we'll see.