Hack around static Nix store timestamps
My RSS feed never appears to update because it's returning a
Last-Modified header of 1980-01-01...

I want to move this to part of the build process I think. The site build
would no longer be deterministic then, which is a shame, but better than
having to manually update this BUILD_VERSION all the time.