Trust myself to update the store
This is needed for the new deployment method. It was a fun old time
fixing this without creating a new VPS. I had to move the existing
nix.conf out of the way, then create a new one I could edit, then add
wheel, then restart the nix-daemon. Then away we go.